adds set_gdsii_timestamp function and test. moves hashing tests to tests/gdslibrary.py
